+++ client/src/widgets/index.ts
@@ -1,0 +1,2 @@
+export * from "./dimension-widget-interface"
+export * from "./sensemake-resource"