+++ tests/src/index.ts
@@ -7,10 +7,10 @@ import range from './sensemaker_dna/sensemaker/range';
 import dashboard from './sensemaker_dna/sensemaker/dashboard';
 import agent from './sensemaker_dna/sensemaker/agent';
 range();
-neighbourhood();
-applet_config();
-sensemaker_config();
-sensemaker_community_activator();
-dashboard();
-agent();
+// neighbourhood();
+// applet_config();
+// sensemaker_config();
+// sensemaker_community_activator();
+// dashboard();
+// agent();